requirements.txt
flask
flask-cors
PyJWT
mysql-connector-python
werkzeug
